# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZK-DPP Royalty Protocol API
  description: |
    API documentation for the Zero-Knowledge Digital Product Passport (DPP) Royalty Protocol.

    This protocol enables privacy-preserving compliance verification for supply chain data,
    allowing suppliers to prove compliance without exposing trade secrets.

    ## Services Overview

    The system consists of three microservices:

    | Service | Port | Purpose |
    |---------|------|---------|
    | verify-gateway | 3001 | ZK proof verification and receipt issuance |
    | dpp-builder | 3002 | Product management and DPP view composition |
    | metering-billing | 3003 | Usage tracking and settlement management |

    ## Authentication

    Most endpoints require JWT authentication via Keycloak. Include the token in the Authorization header:

    ```
    Authorization: Bearer <token>
    ```

    ## Access Levels

    DPP views are tiered by access level:
    - **PUBLIC**: No authentication required
    - **LEGIT_INTEREST**: Requires `auditor` or `authority` role
    - **AUTHORITY**: Requires `authority` role
  version: 0.1.0
  license:
    name: Apache-2.0 OR MIT
    url: https://github.com/hadijannat/zk-dpp-royalty-protocol/blob/main/LICENSE
  contact:
    name: ZK-DPP Team
    url: https://github.com/hadijannat/zk-dpp-royalty-protocol
servers:
  - url: http://localhost:3001
    description: Verify Gateway (local development)
  - url: http://localhost:3002
...
  - url: http://localhost:3003
...
paths:
...
  /dpp/{id}/view/public:
    get:
      tags: [DPP Views]
      summary: Get public DPP view
      description: |
        Returns the public view of a Digital Product Passport.
        No authentication required.

        Contains: basic product info, category, public metadata
      operationId: getDppPublicView
      servers:
        - url: http://localhost:3002
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product UUID
      responses:
        '200':
          description: Public DPP view
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  dpp:
                    $ref: '#/components/schemas/DppPublicView'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dpp/{id}/view/legit-interest:
    get:
      tags: [DPP Views]
      summary: Get legitimate interest DPP view
      description: |
        Returns the legitimate interest view of a Digital Product Passport.
        Requires authentication with `auditor` or `authority` role.

        Contains: public info + verified predicate results (boolean pass/fail, no raw values)
      operationId: getDppLegitInterestView
      servers:
        - url: http://localhost:3002
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product UUID
      responses:
        '200':
          description: Legitimate interest DPP view
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  dpp:
                    $ref: '#/components/schemas/DppLegitInterestView'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dpp/{id}/view/authority:
    get:
      tags: [DPP Views]
      summary: Get authority DPP view
      description: |
        Returns the full authority view of a Digital Product Passport.
        Requires authentication with `authority` role.

        Contains: all info + raw values + full audit trail
      operationId: getDppAuthorityView
      servers:
        - url: http://localhost:3002
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product UUID
      responses:
        '200':
          description: Authority DPP view
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  dpp:
                    $ref: '#/components/schemas/DppAuthorityView'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===========================================
  # METERING & BILLING (Port 3003)
  # ===========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        success:
          type: boolean
          default: false
    AuthError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum: [AUTH_REQUIRED, INSUFFICIENT_ROLE]
...
    SupplierLink:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        supplierId:
          type: string
        commitmentRoot:
          type: string
        supplierPublicKey:
          type: string
        linkedAt:
          type: string
          format: date-time
...
    DppPublicView:
      type: object
      properties:
        productId:
          type: string
        sku:
          type: string
        name:
          type: string
        category:
          type: string
        accessLevel:
          type: string
          enum: [PUBLIC]
    DppLegitInterestView:
      allOf:
        - $ref: '#/components/schemas/DppPublicView'
        - type: object
          properties:
            accessLevel:
              type: string
              enum: [LEGIT_INTEREST]
            verifiedPredicates:
              type: array
              items:
                type: object
                properties:
                  predicateId:
                    type: string
                  passed:
                    type: boolean
                  verifiedAt:
                    type: string
                    format: date-time
    DppAuthorityView:
      allOf:
        - $ref: '#/components/schemas/DppLegitInterestView'
        - type: object
          properties:
            accessLevel:
              type: string
              enum: [AUTHORITY]
            supplierLinks:
              type: array
              items:
                $ref: '#/components/schemas/SupplierLink'
            auditTrail:
              type: array
              items:
                type: object
                properties:
                  action:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  actor:
                    type: string
...